+++ devices/pic24fj256gb106/boot_user.c
@@ -239,6 +239,6 @@ void writeMax(uint32_t address, uint32_t* progData){
 	}
 }
 
-void startApp(uint16_t applicationAddress){
+void startApp(uint32_t applicationAddress){
 	__asm__("goto w0");
 }
